requirements.txt
absl-py=2.1.0
aiohappyeyeballs=2.4.4 # May need to check conda-forge availability or if default channel has it
aiohttp=3.11.10
aiosignal=1.2.0
alsa-lib=1.2.13
anyio=4.8.0
argon2-cffi=23.1.0
argon2-cffi-bindings=21.2.0
arrow=1.3.0
asttokens=3.0.0
async-lru=2.0.4
attrs=25.1.0
aws-c-auth=0.8.6
aws-c-cal=0.8.7
aws-c-common=0.12.0
aws-c-compression=0.3.1
aws-c-event-stream=0.5.4
aws-c-http=0.9.4
aws-c-io=0.17.0
aws-c-mqtt=0.12.2
aws-c-s3=0.7.13
aws-c-sdkutils=0.2.3
aws-checksums=0.2.3
aws-crt-cpp=0.31.0
aws-sdk-cpp=1.11.510
azure-core-cpp=1.14.0
azure-identity-cpp=1.10.0
azure-storage-blobs-cpp=12.13.0
azure-storage-common-cpp=12.8.0
azure-storage-files-datalake-cpp=12.12.0
babel=2.17.0
beautifulsoup4=4.13.3
bleach=6.2.0
bleach-with-css=6.2.0
bottleneck=1.4.2
brotli=1.1.0
brotli-bin=1.1.0
brotli-python=1.1.0
bzip2=1.0.8
c-ares=1.34.4
ca-certificates=2025.2.25 # May need to check conda-forge availability or if default channel has it
cached-property=1.5.2
cached_property=1.5.2
cairo=1.18.4
certifi=2025.4.26
cffi=1.17.1
charset-normalizer=3.4.1
cloudpickle=3.0.0
colorama=0.4.6
comm=0.2.2
contourpy=1.3.1
# cpython=3.13.2 # Python version managed by conda create
cycler=0.12.1
cyrus-sasl=2.1.27
dataclasses=0.8
datasets=3.6.0
dbus=1.13.6
debugpy=1.8.13
decorator=5.2.1
defusedxml=0.7.1
dill=0.3.8
double-conversion=3.3.1
exceptiongroup=1.2.2
executing=2.1.0
expat=2.6.4
filelock=3.17.0
font-ttf-dejavu-sans-mono=2.37
font-ttf-inconsolata=3.000
font-ttf-source-code-pro=2.038
font-ttf-ubuntu=0.83
fontconfig=2.15.0
fonts-conda-ecosystem=1
fonts-conda-forge=1
fonttools=4.56.0
fqdn=1.5.1
freetype=2.12.1
frozenlist=1.5.0
fsspec=2025.3.0
gflags=2.2.2 # May need to check conda-forge availability or if default channel has it
giflib=5.2.2
glog=0.7.1
gmp=6.3.0
gmpy2=2.1.5
graphite2=1.3.13
grpcio=1.67.1
gymnasium=0.26.3
gymnasium-notices=0.0.1
h11=0.14.0
h2=4.2.0
harfbuzz=10.4.0
hpack=4.1.0
httpcore=1.0.7
httpx=0.28.1
huggingface_hub=0.31.2
hyperframe=6.1.0
icu=75.1
idna=3.10
importlib-metadata=8.6.1
importlib_resources=6.5.2
ipykernel=6.29.5
ipython=9.0.2
ipython_pygments_lexers=1.1.1
isoduration=20.11.0
jedi=0.19.2
jinja2=3.1.6
joblib=1.4.2
json5=0.10.0
jsonpointer=3.0.0
jsonschema=4.23.0
jsonschema-specifications=2024.10.1
jsonschema-with-format-nongpl=4.23.0
jupyter-lsp=2.2.5
jupyter_client=8.6.3
jupyter_core=5.7.2
jupyter_events=0.12.0
jupyter_server=2.15.0
jupyter_server_terminals=0.5.3
jupyterlab=4.3.5
jupyterlab_pygments=0.3.0
jupyterlab_server=2.27.3
keyutils=1.6.1
kiwisolver=1.4.7
krb5=1.21.3
lcms2=2.17
lerc=4.0.0
libabseil=20240722.0
libarrow=19.0.1
libarrow-acero=19.0.1
libarrow-dataset=19.0.1
libarrow-substrait=19.0.1
libblas=3.9.0
libbrotlicommon=1.1.0
libbrotlidec=1.1.0
libbrotlienc=1.1.0
libcblas=3.9.0
libclang-cpp19.1=19.1.7
libclang13=19.1.7
libcrc32c=1.1.2 # May need to check conda-forge availability or if default channel has it
libcups=2.3.3
libcurl=8.13.0
libdeflate=1.23
libdrm=2.4.124
libedit=3.1.20250104
libegl=1.7.0
libev=4.33 # May need to check conda-forge availability or if default channel has it
libevent=2.1.12 # May need to check conda-forge availability or if default channel has it
libexpat=2.6.4
libffi=3.4.6
libgcc=14.2.0
libgcc-ng=14.2.0
libgfortran=14.2.0
libgfortran5=14.2.0
libgl=1.7.0
libglib=2.82.2
libglvnd=1.7.0
libglx=1.7.0
libgoogle-cloud=2.36.0
libgoogle-cloud-storage=2.36.0
libgrpc=1.67.1
libhwloc=2.11.2
libiconv=1.18
libjpeg-turbo=3.0.0
liblapack=3.9.0
libllvm19=19.1.7
liblzma=5.6.4
libmpdec=4.0.0
libnghttp2=1.64.0
libntlm=1.8
libopengl=1.7.0
libopentelemetry-cpp=1.18.0
libopentelemetry-cpp-headers=1.18.0
libparquet=19.0.1
libpciaccess=0.18
libpng=1.6.47
libpq=17.4
libprotobuf=5.28.3
libre2-11=2024.07.02
libsodium=1.0.20
libsqlite=3.49.1
libssh2=1.11.1
libstdcxx=14.2.0
libstdcxx-ng=14.2.0
libthrift=0.21.0
libtiff=4.7.0
libtorch=2.6.0
libutf8proc=2.10.0
libuuid=2.38.1
libuv=1.50.0
libwebp-base=1.5.0
libxcb=1.17.0
libxkbcommon=1.8.0
libxml2=2.13.6
libxslt=1.1.39
libzlib=1.3.1
llvm-openmp=19.1.7
lz4-c=1.10.0
markdown=3.6
markupsafe=3.0.2
matplotlib=3.10.1
matplotlib-base=3.10.1
matplotlib-inline=0.1.7
mistune=3.1.2
mkl=2024.2.2
mpc=1.3.1
mpfr=4.2.1
mpmath=1.3.0
multidict=6.1.0
multiprocess=0.70.15
munkres=1.1.4
mysql-common=9.0.1
mysql-libs=9.0.1
nbclient=0.10.2
nbconvert-core=7.16.6
nbformat=5.10.4
ncurses=6.5
nest-asyncio=1.6.0
networkx=3.4.2
nlohmann_json=3.11.2 # May need to check conda-forge availability or if default channel has it
notebook-shim=0.2.4
numexpr=2.10.2
numpy=2.2.6
openjpeg=2.5.3
openldap=2.6.9
openssl=3.4.1
optree=0.14.1
orc=2.1.1
overrides=7.7.0
packaging=24.2
pandas=2.2.3
pandocfilters=1.5.0
parso=0.8.4
pcre2=10.44
pexpect=4.9.0
pickleshare=0.7.5
pillow=11.1.0
pip=25.0.1
pixman=0.44.2
pkgutil-resolve-name=1.3.10
platformdirs=4.3.6
prometheus-cpp=1.3.0
prometheus_client=0.21.1
prompt-toolkit=3.0.50
propcache=0.3.1 # May need to check conda-forge availability or if default channel has it
protobuf=5.28.3
psutil=7.0.0
pthread-stubs=0.4
ptyprocess=0.7.0
pure_eval=0.2.3
pyarrow=19.0.1
pyarrow-core=19.0.1
pybind11=2.13.6
pybind11-global=2.13.6
pycparser=2.22
pygments=2.19.1
pyparsing=3.2.1
pyside6=6.8.2
pysocks=1.7.1
python-dateutil=2.9.0.post0
python-fastjsonschema=2.21.1
python-json-logger=2.0.7
python-tzdata=2025.2 # May need to check conda-forge availability or if default channel has it
python-xxhash=3.5.0 # May need to check conda-forge availability or if default channel has it
pytorch=2.6.0
pytz=2025.1
pyyaml=6.0.2
pyzmq=26.2.1
qhull=2020.2
qt6-main=6.8.2
re2=2024.07.02
readline=8.2
referencing=0.36.2
regex=2024.11.6
requests=2.32.3
rfc3339-validator=0.1.4
rfc3986-validator=0.1.1
rpds-py=0.23.1
s2n=1.5.14 # May need to check conda-forge availability or if default channel has it
safetensors=0.5.3
scikit-learn=1.6.1
scipy=1.15.2
send2trash=1.8.3
setuptools=75.8.2
six=1.17.0
sleef=3.8
snappy=1.2.1 # May need to check conda-forge availability or if default channel has it
sniffio=1.3.1
soupsieve=2.5
stack_data=0.6.3
standard-imghdr=3.13.0
sympy=1.13.3
tbb=2021.13.0
tensorboard=2.19.0
tensorboard-data-server=0.7.0
terminado=0.18.1
threadpoolctl=3.5.0
tinycss2=1.4.0
tk=8.6.13
tokenizers=0.21.0
tomli=2.2.1
torchvision=0.21.0
tornado=6.4.2
tqdm=4.67.1
traitlets=5.14.3
transformers=4.51.3
types-python-dateutil=2.0.0.20240106 # Version updated for likely availability on conda-forge for older py
typing-extensions=4.12.2
typing_extensions=4.12.2
typing_utils=0.1.0
tzdata=2025a
uri-template=1.3.0
urllib3=2.2.2
wayland=1.23.1
wcwidth=0.2.13
webcolors=24.11.1
webencodings=0.5.1
websocket-client=1.8.0
werkzeug=3.1.3
xcb-util=0.4.1
xcb-util-cursor=0.1.5
xcb-util-image=0.4.0
xcb-util-keysyms=0.4.1
xcb-util-renderutil=0.3.10
xcb-util-wm=0.4.2
xkeyboard-config=2.43
xorg-libice=1.1.2
xorg-libsm=1.2.5
xorg-libx11=1.8.11
xorg-libxau=1.0.12
xorg-libxcomposite=0.4.6
xorg-libxcursor=1.2.3
xorg-libxdamage=1.1.6
xorg-libxdmcp=1.1.5
xorg-libxext=1.3.6
xorg-libxfixes=6.0.1
xorg-libxi=1.8.2
xorg-libxrandr=1.5.4
xorg-libxrender=0.9.12
xorg-libxtst=1.2.5
xorg-libxxf86vm=1.1.6
xxhash=0.8.0 # May need to check conda-forge availability or if default channel has it
yaml=0.2.5
yarl=1.18.0
zeromq=4.3.5
zipp=3.21.0
zlib=1.3.1
zstd=1.5.7
perf